Fixes connect button request on profile gallery.
